+++ src/board/supported/jingcai/BOARD_JINGCAI_ESP32_4848S040C_I_Y_3.h
@@ -345,6 +345,23 @@
     #define ESP_PANEL_BOARD_BACKLIGHT_IO        (38)    // Output GPIO pin number
     #define ESP_PANEL_BOARD_BACKLIGHT_ON_LEVEL  (1)     // Active level, 0: low, 1: high
 
+
+#if ESP_PANEL_BOARD_BACKLIGHT_TYPE == ESP_PANEL_BACKLIGHT_TYPE_PWM_LEDC
+    /**
+     * @brief PWM parameters configuration
+     */
+    #define ESP_PANEL_BOARD_BACKLIGHT_PWM_FREQ_HZ   (1000)  // LEDC timer frequency.
+                                                            // Different backlight driver chips may have different
+                                                            // frequency limits, please refer to the datasheet of
+                                                            // the specific chip.
+                                                            // https://github.com/esp-arduino-libs/ESP32_Display_Panel/issues/188
+
+    #define ESP_PANEL_BOARD_BACKLIGHT_PWM_DUTY_RESOLUTION  (10) // LEDC timer duty resolution.
+                                                                // The frequency and duty resolution of the LEDC timer
+                                                                // need to be properly matched, please refer to:
+                                                                // https://docs.espressif.com/projects/esp-idf/en/latest/esp32/api-reference/peripherals/ledc.html#supported-range-of-frequency-and-duty-resolutions
+#endif
+
 #endif // ESP_PANEL_BOARD_BACKLIGHT_TYPE
 
 /**
@@ -381,7 +398,7 @@
  * 3. Patch version mismatch: No impact on functionality
  */
 #define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MAJOR 1
-#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MINOR 0
+#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MINOR 1
 #define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_PATCH 0
 
 // *INDENT-ON*